% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonGenes.R
\name{getCommonGenes}
\alias{getCommonGenes}
\title{Get common genes}
\usage{
getCommonGenes(InterestedDGs)
}
\arguments{
\item{InterestedDGs}{A dataset containing a list of genes of interest}
}
\value{
A dataset with only the common genes
}
\description{
This function intersects a dataset of unique gene symbols with a dataset of interested genes,
and returns only the genes that are common to both datasets,
representing the genes that are shared between the disease of interest and other diseases.
}
\examples{
# Load interested disease genes from file
gene_list <- read.table("gene_list.txt", header = T, sep = '\t')

# Get common genes with a gene dataset
getCommonGenes(gene_list)
}
