% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommunities.R
\name{getCommunities}
\alias{getCommunities}
\title{Identify modules in a disease network}
\usage{
getCommunities(disease_net)
}
\arguments{
\item{disease_net}{A list of edges representing a disease network}
}
\value{
A list containing two data frames: "nodes" with two columns "name" and "module",
where "name" is the name of the node and "module" is the index of the module to which the node belongs;
and "edges" with three columns "name1", "name2", and "module", where "name1" and "name2" are the names
of the nodes connected by the edge and "module" is the index of the module to which the edge belongs.
Both data frames are saved as text files in the "output_Files" directory with the names "node_Module.txt"
and "edge_Module.txt", respectively.
}
\description{
Identify modules in a disease network
}
\examples{
modules_info <- getCommunities(disease_net)
}
