% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnlargedNet.R
\name{getEnlargedNet}
\alias{getEnlargedNet}
\title{Extracts an enlarged subgraph of a given seed network based on reachability score}
\usage{
getEnlargedNet(seed_genes, reachable_genes_threshold = 0.9)
}
\arguments{
\item{seed_genes}{a data frame containing the seed genes}

\item{reachable_genes_threshold}{a numeric value representing the reachability score threshold for identifying significant genes}
}
\value{
a data frame containing the enlarged subgraph of the seed network
}
\description{
This function takes a seed network, calculates reachability scores of each non-seed gene using
Random Walk with Restart (RWR) algorithm, calculates fold change and p-values between seed and non-seed genes,
identifies significant genes based on the p-value and log2 fold change values, and then extracts an enlarged
subgraph that includes the seed genes and the significant non-seed genes.
The enlargement threshold can be adjusted using the reachable_genes_threshold parameter.
}
\examples{
enlarged_net <- getEnlargedNet(seed_genes, 0.9)
}
