% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitialNet.R
\name{getInitialNet}
\alias{getInitialNet}
\title{Generate a sub-network consisting of the common genes and their interacting partners}
\usage{
getInitialNet(common_genes, outnet_name)
}
\arguments{
\item{common_genes}{A dataset containing a list of common genes}

\item{outnet_name}{The name to be given to the output network files}
}
\value{
A list containing the dataset of seed genes and the sub-network
}
\description{
This function takes in a dataset of common genes and an output network name, and returns a sub-network that includes
only the common genes and their interacting partners from the human protein-protein interaction network. The sub-network
and a dataset of seed genes are also written to separate files in the output_Files folder.
}
\examples{
common_genes <- getCommonGenes(InterestedDGs)
seed_info <- getInitialNet(common_genes, "Initial_Network")
seed_genes <- seed_info$seed_genes
seed_net <- seed_info$seed_net
# The output files are saved in the output_Files folder as "Initial_Network_Genes.txt" and "Initial_Network_Net.txt"
# respectively.
}
