% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTargetModule.R
\name{identifyTargetModule}
\alias{identifyTargetModule}
\title{Identify target module in disease gene network}
\usage{
identifyTargetModule(disease_net)
}
\arguments{
\item{disease_net}{A list of edges representing a disease network}
}
\description{
This function first detects modules in a given disease gene network, then maps targets to genes
in the network. It then identifies the target module, and maps drugs to targets in the target module.
The final result is a data frame of drug-target interactions in the target module.
}
\examples{
# Creating a disease network using createDiseaseNetwork function
disease_net <- createDiseaseNetwork(gene_list, enlarge_network = FALSE)

# Creating a disease network from a text file containing network edges
disease_net <- read.table("disease_net.txt", header = TRUE)

# Identifying the target module in the network using identifyTargetModule function
identifyTargetModule(disease_net)
}
