% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDrugs.R
\name{mapDrugs}
\alias{mapDrugs}
\title{Map drugs to target module}
\usage{
mapDrugs(target_module_info)
}
\arguments{
\item{target_module_info}{list of nodes, edges, and targets for the target module}
}
\description{
Map drugs to target module
}
\examples{
mapDrugs(target_module_info)
}
