% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTargets.R
\name{mapTargets}
\alias{mapTargets}
\title{Identify targets of a gene network using DrugBank and TTD databases}
\usage{
mapTargets(Gene_net)
}
\arguments{
\item{Gene_net}{a data frame representing a gene network, where the first two columns represent
the gene symbols of the source and target nodes, respectively.}
}
\value{
a data frame containing the symbols of the genes in the input network that are also targets in
either DrugBank or TTD databases
}
\description{
Identify targets of a gene network using DrugBank and TTD databases
}
\examples{
targets <- mapTargets(disease_net)
head(targets)
}
