% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRWR.R
\name{runRWR}
\alias{runRWR}
\title{Run Random Walk with Restart (RWR) on a protein-protein interaction network (PPIN)}
\usage{
runRWR(seeds, restart = 0.8, output = FALSE)
}
\arguments{
\item{seeds}{A data frame containing a list of gene symbols to be used as input seeds for RWR.}

\item{restart}{A numeric value indicating the restart probability for RWR (default = 0.8).}

\item{output}{A logical value indicating whether or not to output the RWR results as a text file (default = FALSE).}
}
\value{
A data frame containing the RWR scores for each gene in the PPIN.
}
\description{
This function runs RWR on a PPIN to prioritize genes based on their network proximity to the input seeds.
}
\examples{
# Run RWR with default restart probability and no output file
runRWR(seeds = data.frame("BRCA1", "TP53"))

# Run RWR with custom restart probability and output file
runRWR(seeds = data.frame("BRCA1", "TP53"), restart = 0.7, output = TRUE)
}
